rename(card) : cardType의 파일명을 card로 변경
